add 게시글 좋아요 기능 추가, update 게시글 상세 보기 쿼리 수정

+++ communityJh/src/main/java/kr/co/communityJh/board/controller/api/BoardApiController.java
@@ -1,17 +1,20 @@
 package kr.co.communityJh.board.controller.api;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
-import kr.co.communityJh.auth.AuthUser;
 import kr.co.communityJh.account.dto.AccountRequestDto;
+import kr.co.communityJh.auth.AuthUser;
+import kr.co.communityJh.board.dto.LikeRequestDto;
 import kr.co.communityJh.board.service.BoardService;
+import kr.co.communityJh.exception.CustomException;
+import kr.co.communityJh.exception.ErrorCode;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
 /**
  * @author "jhlee"
@@ -22,31 +25,11 @@ import kr.co.communityJh.board.service.BoardService;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/board")
+@Slf4j
 @Api(tags = "커뮤니티_API문서")
 public class BoardApiController {
 
 	private final BoardService boardService;
-	
-//	/**
-//	 * @return 게시글 목록
-//	 */
-//	@PostMapping("/board")
-//	@ApiOperation(notes = "커뮤니티 게시글 목록을 조회 할 수 있습니다.", value = "목록조회")
-//	public List<Board> boards(Board board) {
-//		System.out.println(board);
-//		service.save(board);
-//		return null;
-//	}
-	
-//	@GetMapping("/{seq}")
-//	@ApiOperation(notes = "커뮤니티 게시글을 조회 할 수 있습니다.", value = "게시글 조회")
-//	@ApiImplicitParams({
-//		@ApiImplicitParam(name = "seq", value = "id", example = "1")
-//		
-//	})
-//	public Board boards(@PathVariable int seq) {
-//		return service.selectList();
-//	}
 
 	/**
 	 * 게시글 삭제
@@ -62,7 +45,26 @@ public class BoardApiController {
 	public ResponseEntity<Object> boardDelete(@PathVariable Long id,
 			@AuthUser AccountRequestDto accountRequestDTO) {
 
+		if (accountRequestDTO == null) {
+			throw new CustomException(ErrorCode.UNAUTHORIZED_USER);
+		}
 		boardService.deleteBoardById(id, accountRequestDTO);
 		return new ResponseEntity<>("success", HttpStatus.OK);
+	}
+
+	@ApiOperation(notes = "게시글 id에 해당하는 게시글의 좋아요 상태를 변경합니다.", value = "종아요 기능" )
+	@ApiImplicitParams({
+			@ApiImplicitParam(name = "id", value = "id", example = "1"),
+	})
+	@PutMapping("/{boardId}/like")
+	public ResponseEntity boardLike(@RequestBody LikeRequestDto likeRequestDto,
+			@AuthUser AccountRequestDto accountRequestDTO) {
+
+		if (accountRequestDTO == null) {
+			throw new CustomException(ErrorCode.UNAUTHORIZED_USER);
+		}
+		likeRequestDto.setAccountId(accountRequestDTO.getId());
+
+		return ResponseEntity.ok(boardService.changeLike(likeRequestDto, accountRequestDTO));
 	}
 }
